requirements.txt
python-telegram-bot==20.3
yfinance
pandas
ta
matplotlib
flask
